keystone2: remove earlyprintk from default env variable
User may add earlyprintk for debug as needed. Usually puts it in
args_all variable and that results in double instance of earlyprint
in the bootargs. This causes earlyprintk code to go in a infinite
loop in earlyprintk code with prints showing something like this.

[    0.000000] Booting Linux on physical CPU 0x0
[    0.000000] Booting Linux on physical CPU 0x0
[    0.000000] Booting Linux on physical CPU 0x0

So remove this from default.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>