keystone: net: add second queue manager configuration
QMSS in Keystone2 K2HK has 2 physical queue managers.
This patch adds the configuration of the second queue
manager found in the QMSS on the K2HK platform.
Calling function of the new APIs is assumed to make sure
that the second queue manager exists before calling
them.  Notice that the same descriptors are used when
configuring the 2nd queue manager, i.e. it is assumed
that only one of the 2 queue managers is active at any
time.  This is the case in u-boot because u-boot closes one
ethernet interface and the corresponding queue manager
before it opens another one when trying to find a connected
ethernet port for boot up purpose.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>