k2e: fix get_ddr_seg_num function
The get_ddr_seg_num() has to return the number of 4KB segments of the
entire DDR3 memory installed to the board. Instead it returned the constant
number of segments for the default 2GB SODIMM. That causes that not entire
DDR3 memory is erased during ECC initialization.

This patch fixes the issue.

Signed-off-by: Vitaly Andrianov <vitalya@ti.com>